vector-store: extract column mod from table
